sdp_message_bot/requirements.txt
python-telegram-bot[callback-data]==22.0
python-dotenv==1.1.0
pytz==2025.2
